Spring boot - update time of click and comment good form of variables

+++ Spring/src/main/java/com/example/webShop/Product/ProductRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends CrudRepository<Product, Long> {
-    Product findById(long id);
+public interface ProductRepository extends CrudRepository<Product, Integer> {
+    Product findById(int id);
 
 }
